requirements.txt
anyio==3.7.0
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
python-telegram-bot==20.3
redis==4.5.5
six==1.16.0
sniffio==1.3.0
spotipy==2.23.0
urllib3==1.26.7
spotdl
python-dotenv~=1.0.0